requirements.txt
python-dotenv
google-generativeai
langchain
PyPDF2
faiss-cpu
streamlit
pdfplumber
langchain-community
google-generativeai
transformers
torch
sentence-transformers

